Cleanup: consolidate tests, fix docs, remove README async section
README:
- Remove async support section (unnecessary on landing page)

Documentation:
- Add S7CommPlus API docs with experimental warning
- Add experimental warning to AsyncClient docs
- Update PLC support matrix for S7CommPlus V1/V2 status

Test consolidation (no test logic changed):
- Merge test_server_coverage.py into test_server.py
- Merge test_partner_coverage.py into test_partner.py
- Merge test_logo_coverage.py into test_logo_client.py
- Merge test_db_coverage.py into test_util.py
- Rename test_s7protocol_coverage.py to test_s7protocol.py

Mypy fixes:
- Widen Row.set_value type to accept date/datetime/timedelta
- Add type annotations in test_s7protocol.py, test_partner.py,
  test_connection.py, test_async_client.py
