+++ snap7/util/db.py
@@ -636,7 +636,7 @@ class Row:
         raise ValueError
 
     def set_value(
-        self, byte_index: Union[str, int], type_: str, value: Union[bool, str, float]
+        self, byte_index: Union[str, int], type_: str, value: Union[bool, str, float, date, datetime, timedelta]
     ) -> Optional[Union[bytearray, memoryview]]:
         """Sets the value for a specific type in the specified byte index.
 
@@ -687,7 +687,7 @@ class Row:
             set_wstring(bytearray_, byte_index, value, max_size_int)
             return None
 
-        if type_ == "REAL":
+        if type_ == "REAL" and isinstance(value, (bool, str, float, int)):
             return set_real(bytearray_, byte_index, value)
 
         if type_ == "LREAL" and isinstance(value, float):
